+++ c10/macros/Export.h
@@ -5,7 +5,7 @@
 #include <c10/macros/cmake_macros.h>
 #endif // C10_USING_CUSTOM_GENERATED_MACROS
 
-#include <torch/standalone/macros/Export.h>
+#include <torch/headeronly/macros/Export.h>
 
 // This one is being used by libtorch.so
 #ifdef CAFFE2_BUILD_MAIN_LIB
