+++ c10/ovrsource_defs.bzl
@@ -74,7 +74,7 @@ def define_c10_ovrsource(name, is_mobile):
             ],
         }),
         exported_deps = [
-            "//xplat/caffe2/torch/standalone:torch_standalone_headers",
+            "//xplat/caffe2/torch/headeronly:torch_headeronly",
             ":ovrsource_c10_cmake_macros.h",
             "//arvr/third-party/gflags:gflags",
             "//third-party/cpuinfo:cpuinfo",
